+++ src/PostsTagger.mts
@@ -8,16 +8,17 @@ import {
 } from "./types/index.mjs";
 import {
   deleteDir,
+  ensureDirExisted,
   getCustomExcerpt,
   getExcerpt,
   getRelativePath,
   getUrlPath,
+  hasExisted,
   readByStream,
   writeByStream,
 } from "./utils.mjs";
 import { PostsPaginator } from "./PostsPaginator.mjs";
 import { PostsCollection } from "./PostsCollection.mjs";
-import { access, writeFile } from "fs/promises";
 
 export class PostsTagger {
   public static get basename() {
@@ -112,33 +113,42 @@ export class PostsTagger {
     this.collect(rawItem);
   }
 
-  private async paginate(posts: PostItem[], tag: string) {
+  private getPaginator() {
+    let paginator;
     const { itemsPerPage } = this.options;
     if (itemsPerPage) {
-      const paginator = new PostsPaginator({
+      paginator = new PostsPaginator({
         itemsPerPage,
-        outputDir: this.options.outputDir!,
+        outputDir: this.outputDir,
         baseUrl: this.baseUrl,
-        prefix: `${PostsCollection.basename}_${tag}`,
       });
-      // await paginator.clean();
-      return paginator.start(posts);
     }
-    return [];
+    return paginator;
+  }
+
+  public async preprocess() {
+    return ensureDirExisted(this.outputDir);
   }
 
   public async save() {
     try {
-      // Convert data
-      const rawTags = this.tagsMap.entries();
-      const tags: PostTagItem[] = [];
-      for (const [tag, posts] of rawTags) {
-        tags.push({
+      // Prepare to paginate
+      const paginator = this.getPaginator();
+      await paginator?.clean();
+      await paginator?.preprocess();
+
+      // Convert data and paginate.
+      const tagsEntry = Array.from(this.tagsMap.entries());
+      const works = tagsEntry.map(async ([tag, posts]) => {
+        const prefix = `${PostsCollection.basename}_${tag}`;
+        const postsPages = await paginator?.process(posts, prefix);
+        return {
           tag,
           posts,
-          postsPages: await this.paginate(posts, tag),
-        });
-      }
+          postsPages,
+        };
+      });
+      const tags: PostTagItem[] = await Promise.all(works);
       const tagsJson = JSON.stringify(tags, null, 0);
 
       // Save
@@ -149,7 +159,9 @@ export class PostsTagger {
   }
 
   public async clean() {
-    await deleteDir(this.outputDir);
+    if (await hasExisted(this.outputDir)) {
+      await deleteDir(this.outputDir);
+    }
   }
 
   public get outputDir() {
@@ -159,19 +171,14 @@ export class PostsTagger {
   public async init(): Promise<void> {
     try {
       const defaultData: PostTagItem[] = [];
-      await writeFile(this.path, JSON.stringify(defaultData, null, 0));
+      await writeByStream(this.path, JSON.stringify(defaultData, null, 0));
     } catch (error) {
       console.error(`Failed create ${PostsTagger.filename}`, error);
     }
   }
 
   public async hasExisted(): Promise<boolean> {
-    try {
-      await access(this.path);
-    } catch {
-      return false;
-    }
-    return true;
+    return hasExisted(this.path);
   }
 
   public async load(): Promise<void> {
